Moving pkg/kubectl/util to staging
