+++ pkg/kubectl/BUILD
@@ -18,7 +18,6 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/kubectl/scheme:go_default_library",
-        "//pkg/kubectl/util:go_default_library",
         "//staging/src/k8s.io/api/apps/v1:go_default_library",
         "//staging/src/k8s.io/api/autoscaling/v1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
@@ -40,6 +39,7 @@ go_test(
         "//staging/src/k8s.io/client-go/scale:go_default_library",
         "//staging/src/k8s.io/client-go/scale/fake:go_default_library",
         "//staging/src/k8s.io/client-go/testing:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
     ],
 )
 
@@ -61,7 +61,9 @@ go_library(
         "//pkg/kubectl/apps:go_default_library",
         "//pkg/kubectl/describe/versioned:go_default_library",
         "//pkg/kubectl/scheme:go_default_library",
-        "//pkg/kubectl/util:go_default_library",
+        "//pkg/kubectl/util/deployment:go_default_library",
+        "//pkg/kubectl/util/podutils:go_default_library",
+        "//pkg/kubectl/util/slice:go_default_library",
         "//staging/src/k8s.io/api/apps/v1:go_default_library",
         "//staging/src/k8s.io/api/autoscaling/v1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
@@ -85,9 +87,7 @@ go_library(
         "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//staging/src/k8s.io/client-go/scale:go_default_library",
         "//staging/src/k8s.io/client-go/util/retry:go_default_library",
-        "//staging/src/k8s.io/kubectl/pkg/util/deployment:go_default_library",
-        "//staging/src/k8s.io/kubectl/pkg/util/podutils:go_default_library",
-        "//staging/src/k8s.io/kubectl/pkg/util/slice:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
         "//vendor/k8s.io/utils/integer:go_default_library",
         "//vendor/k8s.io/utils/pointer:go_default_library",
     ],
@@ -116,7 +116,23 @@ filegroup(
         "//pkg/kubectl/polymorphichelpers:all-srcs",
         "//pkg/kubectl/proxy:all-srcs",
         "//pkg/kubectl/scheme:all-srcs",
-        "//pkg/kubectl/util:all-srcs",
+        "//pkg/kubectl/util/certificate:all-srcs",
+        "//pkg/kubectl/util/deployment:all-srcs",
+        "//pkg/kubectl/util/event:all-srcs",
+        "//pkg/kubectl/util/fieldpath:all-srcs",
+        "//pkg/kubectl/util/hash:all-srcs",
+        "//pkg/kubectl/util/i18n:all-srcs",
+        "//pkg/kubectl/util/interrupt:all-srcs",
+        "//pkg/kubectl/util/logs:all-srcs",
+        "//pkg/kubectl/util/podutils:all-srcs",
+        "//pkg/kubectl/util/printers:all-srcs",
+        "//pkg/kubectl/util/qos:all-srcs",
+        "//pkg/kubectl/util/rbac:all-srcs",
+        "//pkg/kubectl/util/resource:all-srcs",
+        "//pkg/kubectl/util/slice:all-srcs",
+        "//pkg/kubectl/util/storage:all-srcs",
+        "//pkg/kubectl/util/templates:all-srcs",
+        "//pkg/kubectl/util/term:all-srcs",
         "//pkg/kubectl/validation:all-srcs",
         "//pkg/kubectl/version:all-srcs",
     ],
