+++ pkg/kubectl/cmd/portforward/BUILD
@@ -8,7 +8,6 @@ go_library(
     deps = [
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubectl/polymorphichelpers:go_default_library",
-        "//pkg/kubectl/util:go_default_library",
         "//pkg/kubectl/util/i18n:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -18,7 +17,7 @@ go_library(
         "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//staging/src/k8s.io/client-go/tools/portforward:go_default_library",
         "//staging/src/k8s.io/client-go/transport/spdy:go_default_library",
-        "//staging/src/k8s.io/kubectl/pkg/util/templates:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
     ],
 )
