+++ pkg/kubectl/cmd/rollingupdate/BUILD
@@ -9,7 +9,6 @@ go_library(
         "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubectl/scheme:go_default_library",
-        "//pkg/kubectl/util:go_default_library",
         "//pkg/kubectl/util/i18n:go_default_library",
         "//pkg/kubectl/validation:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
@@ -22,7 +21,7 @@ go_library(
         "//staging/src/k8s.io/cli-runtime/pkg/resource:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//staging/src/k8s.io/client-go/scale:go_default_library",
-        "//staging/src/k8s.io/kubectl/pkg/util/templates:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
