+++ pkg/kubectl/rolling_updater_test.go
@@ -40,8 +40,8 @@ import (
 	restclient "k8s.io/client-go/rest"
 	manualfake "k8s.io/client-go/rest/fake"
 	testcore "k8s.io/client-go/testing"
+	"k8s.io/kubectl/pkg/util"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
-	"k8s.io/kubernetes/pkg/kubectl/util"
 )
 
 func oldRc(replicas int, original int) *corev1.ReplicationController {
